Move LiteralNameHelper to FirBuilder.{h,cpp}
